[GFC][Cleanup] Move GridLayoutConstraints and AxisConstraints to their own files https://bugs.webkit.org/show_bug.cgi?id=308659 rdar://171190740
Reviewed by Brandon Stewart.

These structs currently live in GridFormattingContext but they not need
to. We can move them into their own file which keeps the rest of the
code in GridFormattingContext more focused. Also, we can move the
FreeSpaceScenario enum from its own file to the new AxisConstraints one
since it is tied so closely with that.

Canonical link: https://commits.webkit.org/308463@main
