--- Source/WebCore/layout/formattingContexts/grid/FreeSpaceScenario.h
@@ -1,47 +1,0 @@
-/*
- * Copyright (C) 2026 Apple Inc. All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- * 1. Redistributions of source code must retain the above copyright
- *    notice, this list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright
- *    notice, this list of conditions and the following disclaimer in the
- *    documentation and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY APPLE INC. AND ITS CONTRIBUTORS ``AS IS''
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
- * THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
- * PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL APPLE INC. OR ITS CONTRIBUTORS
- * BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
- * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
- * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
- * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF
- * THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-#pragma once
-
-namespace WebCore {
-namespace Layout {
-// https://drafts.csswg.org/css-grid-1/#algo-flex-tracks
-// Enumerate the three free space scenarios for flex track sizing (spec §11.7).
-enum class FreeSpaceScenario : uint8_t {
-    // Scenario 1: Free space is zero or sizing under min-content constraint.
-    // Flex tracks resolve to zero.
-    MinContent,
-
-    // Scenario 2: Free space is a definite length.
-    // Flex tracks distribute available space using findSizeOfFr().
-    Definite,
-
-    // Scenario 3: Free space is infinite or sizing under max-content constraint.
-    // Flex tracks compute from max-content contributions.
-    MaxContent
-};
-
-} // namespace Layout
-} // namespace WebCore
